+++ sdk/hybridconnectivity/azure-mgmt-hybridconnectivity/generated_tests/test_hybrid_connectivity_mgmt_solution_configurations_operations.py
@@ -70,14 +70,7 @@ class TestHybridConnectivityMgmtSolutionConfigurationsOperations(AzureMgmtRecord
             properties={
                 "id": "str",
                 "name": "str",
-                "properties": {
-                    "solutionType": "str",
-                    "lastSyncTime": "2020-02-20 00:00:00",
-                    "provisioningState": "str",
-                    "solutionSettings": {},
-                    "status": "str",
-                    "statusDetails": "str",
-                },
+                "properties": {"solutionSettings": {}, "solutionType": "str"},
                 "systemData": {
                     "createdAt": "2020-02-20 00:00:00",
                     "createdBy": "str",